+++ dspy/dsp/utils/__init__.py
@@ -1,4 +1,3 @@
 from dspy.dsp.utils.dpr import *
-from dspy.dsp.utils.metrics import *
 from dspy.dsp.utils.settings import *
 from dspy.dsp.utils.utils import *
